DATA STRUCTURE TESTS
We started working on data structure test. Theses tests are focused on run residue computation using different data structures and focusing on performance and timing. First, we started with kdtree, octree, compressed octree and trihash.

- Kdtree -> Working. ANN implementation
- Octree -> Working. PCL implementation
- Trihash -> Working. This is the LazerScan code (Yago).
- Compressed Octree -> To Do.

After the inclusion of these datastructures, some test will be planed in order to find the best method for residue computation.